Moved DOM Menu into its own module
